fixes viewport height issue in Chrome mobile
In Chrome Mobile the buttons in the bottom of the modal disappear when the address bar is visible. This is because in Chrome Mobile the address bar is calculated into the viewport height. This PR fixes the issue by using percentage instead of viewport height to make the modal fullscreen